project_root/requirements.txt
Flask
pydub
SpeechRecognition
openai
python-dotenv
apikey
st_audio_recorder
```